Add notes for future perusers of the branch
A bench of common tests shows this module is 33% slower than its currently
active counterpart. The several hot spots that can be optimized also exist in
some form in the "old" module and can be optimized there as well, resulting
in a null performance gain for this version of the module.

I don't see a way to gain 33% improvement from other code paths, so I'm
giving up on this branch and am OK with it being deleted entirely after
the tears over the lost time spent dry out.